backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
gtfs-realtime-bindings==1.0.0
requests==2.32.3
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.6
openai>=1.40.0,<2.0.0
